+++ ouds_core/lib/components/checkbox/ouds_checkbox_item.dart
@@ -15,6 +15,7 @@ library;
 import 'package:flutter/material.dart';
 import 'package:ouds_core/components/checkbox/ouds_checkbox.dart';
 import 'package:ouds_core/components/control/ouds_control_item.dart';
+import 'package:ouds_core/components/control/ouds_control_item_icon.dart';
 
 ///
 /// [OUDS Checkbox design guidelines](https://r.orange.fr/r/S-ouds-doc-checkbox)
@@ -72,7 +73,7 @@ import 'package:ouds_core/components/control/ouds_control_item.dart';
 ///   helperTitle: 'Helper text',
 ///   reversed: false,
 ///   readOnly: false,
-///   icon: 'assets/ic_heart/svg',
+///   icon: OudsControlItemIcon('assets/ic_heart/svg'),
 ///   isError: false,
 ///   divider: true,
 ///   onChanged: (bool? value) {
@@ -90,7 +91,7 @@ class OudsCheckboxItem extends StatelessWidget {
   final ValueChanged<bool?>? onChanged;
   final String title;
   final String? helperTitle;
-  final String? icon;
+  final OudsControlItemIcon? icon;
   final bool reversed;
   final bool readOnly;
   final bool isError;