+++ ouds_core/lib/components/switch/ouds_switch_item.dart
@@ -14,6 +14,7 @@ library;
 
 import 'package:flutter/material.dart';
 import 'package:ouds_core/components/control/ouds_control_item.dart';
+import 'package:ouds_core/components/control/ouds_control_item_icon.dart';
 import 'package:ouds_core/components/switch/ouds_switch.dart';
 import 'package:ouds_core/l10n/gen/ouds_localizations.dart';
 
@@ -78,7 +79,7 @@ class OudsSwitchButtonItem extends StatelessWidget {
   final ValueChanged<bool>? onChanged;
   final String title;
   final String? helperTitle;
-  final String? icon;
+  final OudsControlItemIcon? icon;
   final bool reversed;
   final bool readOnly;
   final bool isError;